move EqSearch to searches folder and adjust packages
